Code-quality fixes: getMemSize() overflow, col_file.cpp, Phase 0 static-analysis findings (#243)
* Widen Object::getMemSize() and overrides to size_t

Image::getMemSize() computed its size correctly but returned int, so
a true size beyond INT_MAX (reachable for a sufficiently large
in-memory image) silently truncated/wrapped to an arbitrary, possibly
negative value. Widen the whole getMemSize() chain (Object, Image,
Layer/LayerImage/LayerFolder, Sprite, Mask, Cel, CelData) from int to
size_t so undo-size accounting and the sprite properties dialog can't
observe a wrapped value.

Fixes #233

* Check fseek() and widen size math in load_col_file()

fseek() return values were unchecked (only ftell()'s failure case was
handled), so a failed seek could leave the file position somewhere
unexpected before the subsequent reads. Separately, `size - 8` was fed
through std::div() as an int, truncating for files >= 2 GiB before the
color count is derived from it.

Check both fseek() calls and compute the quotient/remainder directly
on size_t instead of through the int-only std::div().

Fixes #236

* Clean up Phase 0 static-analysis findings in app/file

Mechanical cleanup of leftover low-priority cppcheck/clang-tidy
findings from the #219 security audit:

- ase_format.cpp: drop the unused `name` local when skipping a
  palette-chunk color entry's name.
- bmp_format.cpp: read_win_bminfoheader() populated several
  WINBMPINFOHEADER fields (biPlanes, biSizeImage, biXPelsPerMeter,
  biYPelsPerMeter, biClrUsed, biClrImportant) that were never
  consulted after loading; read past them without storing instead.
- jpeg_format.cpp: JpegFormat::onSave() set error_mgr::fop but never
  installed the custom error_exit/output_message handlers or a
  setjmp() point (unlike onLoad()), so the assignment was dead and a
  libjpeg fatal error during compression fell through to the default
  handler's exit(), killing the whole process instead of failing the
  save. Wire onSave() up the same way onLoad() already is.
- png_format.cpp: document why the PNG_MAX_PALETTE_LENGTH == 256
  build-time assumption matters.

Fixes #237

+++ src/doc/cel_data.h
@@ -35,7 +35,7 @@ public:
   void setPosition(const gfx::Point& pos) { m_position = pos; }
   void setOpacity(int opacity) { m_opacity = opacity; }
 
-  virtual int getMemSize() const override
+  virtual size_t getMemSize() const override
   {
     ASSERT(m_image);
     // cppcheck-suppress nullPointerRedundantCheck
